Large scale configurable notification system
Implementing scheduler service that will listen network folder and
update the same in SOLR for indexing it.
Another service implemented
using restful will look for subscriber information that will handle
subscriber's input and store it. Notification service will fetch stored
subscriber data along with its notification condition and it will be
passed as query to solr with time, if any recent update is present in
solr, then it will be notified to according subscriber.